+++ src/components/MapSection.tsx
@@ -61,13 +61,13 @@ export default function MapSection() {
 
     return (
         <div className="relative w-screen h-screen">
-            {/* 우측 상단 마이페이지 버튼 */}
+            {/* 우측 상단 로그인 버튼 */}
             <div className="absolute top-6 right-6 z-30">
                 <button
                     className="bg-white shadow-md px-4 py-2 text-indigo-500 font-semibold hover:bg-indigo-500 hover:text-white transition"
                     onClick={() => window.fullpage_api?.moveSlideRight()}
                 >
-                    마이페이지 →
+                    로그인 →
                 </button>
             </div>
