+++ src/index.css
@@ -77,3 +77,26 @@ button:focus-visible {
     background-color: #f9f9f9;
   }
 }
+
+/* 스크롤바 전체 영역 */
+::-webkit-scrollbar {
+  width: 8px; /* 세로 스크롤바 너비 */
+  height: 8px; /* 가로 스크롤바 높이 */
+}
+
+/* 스크롤바 트랙 (배경) */
+::-webkit-scrollbar-track {
+  background: #f0f0f0; /* 트랙 배경색 */
+  border-radius: 4px;
+}
+
+/* 스크롤바 핸들 (움직이는 부분) */
+::-webkit-scrollbar-thumb {
+  background: #888; /* 핸들 색상 */
+  border-radius: 4px;
+}
+
+/* 스크롤바 핸들에 호버 효과 추가 */
+::-webkit-scrollbar-thumb:hover {
+  background: #555; /* 호버 시 핸들 색상 */
+}
